integration: move vector tests to cql_collections.rs
Vector type is definitely more about a CQL collection than about
the Session.
